Rush Help
RushHelp is an app for windows mobile. The motivation behind building
this app was many times in emergencies like accidents or pregnancies we
don’t get help in time and a little delay of few minutes can even cost
us to lose a life. But what if if we get help from the people around us.
The users of this help within a certain geographical range will get a
notification of the victim’s request and if they accept to help the
victim then they will be given the location of victims with the route
navigated on map. There are several other services also included like
Blood Donation and helping the old people.